refactored navbar components to implement smooth scrolling between sections.
