+++ ramses-portfolio/src/Components/contactMe.js
@@ -5,7 +5,7 @@ import Resume from "../images/Resume V.3 Ramses Bermudez.pdf"
 
 function ContactMe() {
   return (
-    <div id="cta" className="Section">
+    <div id="Contact Me" className="Section">
       <h2>Contact Me</h2>
       <h3>
         Ready to join your team and collaborate in the future of software
